Feat : 스터디 메인 컴포넌트 분리
