+++ src/containers/StudyPage/StudyDetail/StudyDetail.jsx
@@ -16,7 +16,7 @@ const THEME = {
 const StudyDetail = ({ ...props }) => {
   const queryClient = useQueryClient();
   const { data } = GetStudyDetail();
-  const { titleData, commentData, contentData, tocItem } =
+  const { titleData, commentData, contentData, tocItem, sidebarData } =
     convertToStudyDetail(mockdata);
 
   const mutation = useMutation("postStudyDetailComment", postComment);
@@ -57,8 +57,7 @@ const StudyDetail = ({ ...props }) => {
           <CommentList data={commentData} {...props} />
         </div>
       </Layout>
-
-      <StudySideBar {...props} {...sidbarargs} />
+      <StudySideBar {...props} data={sidebarData} />
     </>
   );
 };
@@ -69,23 +68,12 @@ StudyDetail.defaultProps = {
 
 export default StudyDetail;
 
-const sidbarargs = {
-  data: {
-    type: "알고리즘",
-    target: "-",
-    people: "4",
-    location: "온라인",
-  },
-
-  badges: ["JavaScript", "TypeScript", "Vue.js", "React", "Redux", "Svelte"],
-};
-
 const mockdata = {
   studyId: 2,
   writerInfo: {
     nickname: "su",
-    ordinal: 0,
-    campus: null,
+    ordinal: 1,
+    campus: "서울",
     authCheck: 0,
   },
   comments: [
@@ -94,8 +82,8 @@ const mockdata = {
       content: "온라인으로는 진행을 안 하시는 건가요?",
       writerInfo: {
         nickname: "su",
-        ordinal: 0,
-        campus: null,
+        ordinal: 1,
+        campus: "서울",
         authCheck: 0,
       },
       createdDate: "2022-01-30T21:22:30",
@@ -106,8 +94,8 @@ const mockdata = {
       content: "너무 참여하고 싶어요!",
       writerInfo: {
         nickname: "su",
-        ordinal: 0,
-        campus: null,
+        ordinal: 1,
+        campus: "서울",
         authCheck: 0,
       },
       createdDate: "2022-01-30T21:23:14",
@@ -118,8 +106,8 @@ const mockdata = {
       content: "너무 참여하고 싶어요!",
       writerInfo: {
         nickname: "su",
-        ordinal: 0,
-        campus: null,
+        ordinal: 1,
+        campus: "서울",
         authCheck: 0,
       },
       createdDate: "2022-01-30T21:23:15",
@@ -130,8 +118,8 @@ const mockdata = {
       content: "너무 참여하고 싶어요!",
       writerInfo: {
         nickname: "su",
-        ordinal: 0,
-        campus: null,
+        ordinal: 1,
+        campus: "서울",
         authCheck: 0,
       },
       createdDate: "2022-01-30T21:23:16",
