+++ src/containers/StudyPage/StudyMainMobile/index.js
@@ -1,0 +1,1 @@
+export { default } from "./StudyMainMobile";
