+++ src/containers/StudyPage/index.js
@@ -8,3 +8,5 @@ export { default as StudyDetail } from "./StudyDetail";
 export { default as StudyDetailMobile } from "./StudyDetailMobile";
 export { default as StudyDetailComment } from "./StudyDetailMobile";
 export { default as StudySideBar } from "./StudySideBar";
+export { default as StudyMain } from "./StudyMain";
+export { default as StudyMainMobile } from "./StudyMainMobile";
