+++ src/pages/StudyPage/StudyPage.jsx
@@ -3,14 +3,11 @@ import styled from "styled-components";
 
 import { useMediaQuery } from "react-responsive";
 
-import { useRecoilValue, useSetRecoilState } from "recoil";
-import { themeState, navTypeState } from "@/recoil/theme";
-
-import { SubNavbar, Input, Selector } from "@/components";
-import { CardGrid, StudyIntro, StudyCardGrid } from "@/containers";
+import { useSetRecoilState } from "recoil";
+import { navTypeState } from "@/recoil/theme";
+import { StudyIntro, StudyMain, StudyMainMobile } from "@/containers";
 
 const StudyPage = () => {
-  const theme = useRecoilValue(themeState);
   const setNavType = useSetRecoilState(navTypeState);
   const [offset, setOffset] = useState(0);
 
@@ -35,70 +32,14 @@ const StudyPage = () => {
     query: "(max-width:980px)",
   });
 
-  const handleCategory = (e) => {
-    console.log(e);
-  };
-
   return (
     <>
       <StudyIntro />
-      {isPc && (
-        <ContentBox>
-          <StickyNavBox>
-            <StickyNav
-              data={categoryData}
-              theme={theme}
-              onClick={handleCategory}
-            />
-          </StickyNavBox>
-          <CardGridBox>
-            <InputBox>
-              <Input
-                theme={theme}
-                icon="search"
-                placeholder="스터디 검색하기"
-              />
-              <Selector
-                theme={theme}
-                placeholder="종류"
-                options={[
-                  { value: "recruit", label: "채용" },
-                  { value: "Algorithm", label: "알고리즘" },
-                  { value: "CS", label: "CS" },
-                  { value: "Project", label: "프로젝트" },
-                ]}
-              />
-              {/* <Selector placeholder="기술 스택" isMulti /> */}
-            </InputBox>
-            <Suspense fallback={<CardGrid theme={theme} isLoading />}>
-              <StudyCardGrid theme={theme} />
-            </Suspense>
-          </CardGridBox>
-        </ContentBox>
-      )}
-      {isMobile && (
-        <>
-          <SubNavMobile
-            data={categoryData}
-            theme={theme}
-            onClick={handleCategory}
-            view="mobile"
-          />
-          <SearchBox>
-            <Input icon="search" placeholder="스터디 검색하기" theme={theme} />
-          </SearchBox>
-          <MobileCardBox>
-            <Suspense fallback={<CardGrid theme={theme} isLoading />}>
-              <StudyCardGrid theme={theme} />
-            </Suspense>
-          </MobileCardBox>
-        </>
-      )}
+      {isPc && <StudyMain categoryData={categoryData} />}
+      {isMobile && <StudyMainMobile categoryData={categoryData} />}
     </>
   );
 };
-
-export default StudyPage;
 
 const categoryData = [
   {
@@ -114,59 +55,4 @@ const categoryData = [
     title: "나의 스터디",
   },
 ];
-
-const ContentBox = styled.div`
-  display: flex;
-  justify-content: center;
-
-  gap: 100px;
-  max-width: 1280px;
-
-  padding: 20px;
-  margin: auto;
-`;
-
-const StickyNavBox = styled.div`
-  padding-top: 86px;
-`;
-
-const StickyNav = styled(SubNavbar)`
-  position: sticky;
-  top: 150px;
-`;
-
-const SubNavMobile = styled(SubNavbar)`
-  padding: 20px 20px 0 20px;
-`;
-
-const CardGridBox = styled.div`
-  width: calc(100% - 320px);
-  padding-top: 80px;
-`;
-
-const MobileCardBox = styled.div`
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  gap: 100px;
-  max-width: 1280px;
-
-  padding: 20px;
-  margin: auto;
-`;
-
-const InputBox = styled.div`
-  display: flex;
-  gap: 0.5rem;
-
-  max-width: 940px;
-  padding-bottom: 2rem;
-
-  > div {
-    flex-grow: 1;
-  }
-`;
-
-const SearchBox = styled.div`
-  padding: 20px;
-`;
+export default StudyPage;
